requirements.txt
# Core dependencies
numpy>=1.26.4
pandas>=2.2.3
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Deep Learning
tensorflow>=2.18.0

# Additional utilities
tqdm>=4.65.0
PyYAML>=6.0
h5py>=3.9.0